Implement proper Lean emission
This commit takes the sketchy code that was already on the branch and
turns it into a proper flexible structure for the emission of Lean code.
It also separates the CLI from the emission library, allowing the tool
to be more flexible.

Please note that this is still a work in progress, and not every feature
of Noir is supported as of yet.