Fix the integer -> field element conversions
Previously these conversions (`uAsField` and `iAsField`) used `sorry` as
we had not clarified the expected semantics of these conversions with
Aztec. We now have clarity on how these should behave, and have
implemented the correct conversion behavior in our builtins. The exact
semantics of these conversions are documented in the doc blocks.

This commit adds the intro theorems for `uAsField` and `iAsField`, and
also removes an incidental example that should not have been in the
codebase. It also adds those theorems to the steps tactic to help make
progress where these are obviously the correct goal.