Fix the Lampe trait resolution tactic (#143)
- The previous `enter_trait ...` tactic was not able to introspect on
  the trait implementations in compound environments of the form
  `def compoundEnv : Lampe.Env := env1 ++ env2`.
- The `try_impls_all ...` tactic was too eager to unify some
  metavariables when used as part of `enter_trait`. The check that the
  tactic was successful did not take into account whether the outside
  call to `callTrait_direct_intro` was able to match on the correct
  traid, and this led to false successes for `try_impls_all`.

This PR solves both by replacing the `enter_trait ...` tactic with a new
`try_impls_all` tactic which combines the call to
`callTrait_direct_intro` and `try_impls_all`. The tactic then unifies
all of the resulting metavariables from the function applications in the
right order for the type inference system to do the rest of the job.

Note that this tactic will likely need to be revisited when we get
around to both #109 and #144.