Prove environment monotonicity for `STHoare` (#161)
In order to compose Noir libraries with their accompanying lampe proofs,
it is necessary to be able to extend proofs performed over some
environment e_1 to any environment e_2 that contains e_1. In other
words, we need to prove that our language semantics are _monotonic_ with
respect to their environment.

To this end, this commit states and proves a theorem for the
monotonicity of `Omni` and then `STHoare` using that. It also contains
proofs of the monotonicity of `TraitResolvable` and `TraitResolution` as
these proved necessary. It also includes a subset relation on `Env`s,
and provides a number of useful lemmas for reasoning about this subset
relation.

Furthermore, it adds a test that demonstrates the correct function of
the monotonicity theorems. This test performs a number of proofs of
functions in isolation, both using traits and not, and shows that those
proofs can be used in proofs that take place in an environment of which
each proof's environment was a subset.

Co-authored-by: Marcin Kostrzewa <marckostrzewa@gmail.com>
Co-authored-by: Matej Penciak <matej.penciak@gmail.com>