Output the new syntax when extracting
Now all extractions will use the new syntax that separates analysis from
generation for clarity and easier adaptability. Note that the Lean
codebase has not yet been updated to handle this new syntax.

This also makes some minor fixes to the analysis to better assist in
correct output of syntax.