Add the stdlib as a project dependency
Please note that the current location and inclusion convention may
change, as it is not final. It is unconditionally included as a git
dependency in any newly-extracted project unless that project is the
standard library itself.

It is included with its current `name-version` string, as should be true
for any extracted dependencies.