Ensure we gather methods on ALL types (#195)
Previously we were using the fact that `impl` blocks for structs were
declared as child modules to gather methods, but this did not work for
impls on builtin types.

This commit changes how we do method discovery for a file in three ways:

- When a module `is_type` (was created to hold impl methods) we ignore
  it at the top level.
- When traversing the entries in a module, we now identify type
  definitions and extract their direct methods, generating them if they
  are defined in the same file.
- When extracting the standard library, we additionally query for
  methods on builtin types, and add them to the current file if they
  were defined in that file.

Additionally, for methods or functions that directly represent calls to
builtins, we no longer generate a stub method and instead generate a
call to the builtin directly where one is encountered.