Add the remaining theorems for `Option` (#211)
This commit adds statements and proofs of the remaining theorems to
accompany the functionality of the `std::option::Option<T>` type. It
also introduces a document that aims to collate useful tips for stating
theorems and proving.

Do note that it does not contain proofs for some of the traits
implemented for `Option` as these will be influenced by the "lawful"
trait initiative, and hence are not worth doing at this stage.

It also adds the `unsafe` mode to both the `steps` and `sl` tactics.
This allows `sl` (and uses thereof embedded in `steps`) to unify from
the entire proof state instead of the state of the current goal.