Add theorems for Option's traits
We were originally going to work on a `Lawful` system for trait
implementations, but in the interests of expediency we are going to
leave that for later (#204). This commit adds basic theorems for the
remaining traits on Option, and also provides some additional ergonomics
for working with traits with `where` bounds.

Co-authored-by: Marcin Kostrzewa <marckostrzewa@gmail.com>