Improve our Prime representation (#219)
While we are mostly agnostic to the prime that is used, we do care that
it is larger than a lower bound (currently two). This is now encoded in
the definition of the `Prime` type to make it possible to prove certain
theorems.

In addition, we now provide a way to easily construct primes using the
n + 1 representation without resorting to awkward definitions of p - 1
to just construct the `Prime`.