Add theorems for conversions (#222)
This commit includes theorems for the instances of `std::convert::Into`
and `std::convert::From` that are part of the standard library. It also
includes added semantics for unsigned integer literals, and more
builtins have been given support in the `steps` tactic.

It also fixes a bug with extracting the wrong trait name in trait
bounds, and updates the standard library extraction as a result.