Add support for deprecated annotations
These can be placed on function and type definitions and are currently
used by the extractor to annotate deprecated things in the extracted
code. Deprecated items are explicitly _not_ added to the environment to
discourage their usage.

This commit also adds a definition for the `arrayAsStrUnchecked`
builtin and a few simple proofs.