Add the remaining `Array` theorems (#233)
This commit also fixes a bug in the backtracking behavior inside the
separation logic and Hoare logic tactics. The addition of the `ReaderT`
monad transformer in the past had changed the `MonadExcept` instance
used by the `try-catch` statement to no longer perform state restoration
on a failure as is done by `TacticM`. This commit re-enables that
behavior.

Co-authored-by: Marcin Kostrzewa <marckostrzewa@gmail.com>